Planning
Many minor changes needed. I need to focus on a lot of things but im close to finishing. 2-3 days until i do something for function to show AS. GetElem Intellisense has a lot of things that i havent considered, but overall very nice touch and progression to it. After Functions i have finished but be carefull with bugs, dont let stackframes and var updating cause me problems in function implementation.